prixod table change

+++ src/controllers/admin-app/prixod.controller.js
@@ -30,6 +30,7 @@ class prixodController {
             data: model
         });
     }
+
     getOne = async (req, res, next) => {
         this.checkValidation(req);
         const model = await prixod_Model.findOne({
@@ -56,8 +57,10 @@ class prixodController {
             data: model
         });
     }
+
     create =  async(req, res, next) => {
        this.checkValidation(req);
+
        const {prixod_table, ...data} = req.body;
        const model = await prixod_Model.create({
             "date_time": Math.floor(new Date().getTime() / 1000),
@@ -65,7 +68,9 @@ class prixodController {
             "umumiy_summa": req.body.umumiy_summa,
             "comment": req.body.comment
        });
+
        this.#prixod_table(model, prixod_table);
+       
        var register = {
          "date_time": Math.floor(new Date().getTime() / 1000),
          "doc_id": model.id,
@@ -75,7 +80,9 @@ class prixodController {
          "place": "Prixod",
          "pastavchik_id": model.pastavchik_id
        }
+       
        await register_supplierModel.create(register);
+       
        res.status(200).send({
             error: false,
             error_code: 200,
@@ -83,47 +90,49 @@ class prixodController {
             data: model
        });
     }
+
     update = async (req, res, next) => {
-       this.checkValidation(req);
-    const model = await prixod_Model.findOne({
-        where:{
-            id: req.params.id
+        this.checkValidation(req);
+        const model = await prixod_Model.findOne({
+            where:{
+                id: req.params.id
+            }
+        });
+
+        const {prixod_table, ...data} = req.body;
+        model.date_time = req.body.date_time;
+        model.pastavchik_id = req.body.pastavchik_id;
+        model.umumiy_summa = req.body.umumiy_summa;
+        model.comment = req.body.comment;
+        model.save();
+        
+        this.#prixod_table(model, prixod_table, false);
+        await this.#deleteRegister_supplier(model.id)
+        
+        var register = {
+            "date_time": Math.floor(new Date().getTime() / 1000),
+            "doc_id": model.id,
+            "summa": model.umumiy_summa,
+            "doc_type": "chiqim",
+            "type": 2,
+            "place": "Prixod",
+            "pastavchik_id": model.pastavchik_id
         }
-    });
-    const {prixod_table, ...data} = req.body;
-    model.date_time = req.body.date_time;
-    model.pastavchik_id = req.body.pastavchik_id;
-    model.umumiy_summa = req.body.umumiy_summa;
-    model.comment = req.body.comment;
-    model.save();
-    this.#prixod_table(model, prixod_table, false);
-    await register_supplierModel.destroy({
-        where:{
-            doc_id: model.id,
-            place: 'Prixod'
-        }
-    })
-    var register = {
-        "date_time": Math.floor(new Date().getTime() / 1000),
-        "doc_id": model.id,
-        "summa": model.umumiy_summa,
-        "doc_type": "chiqim",
-        "type": 2,
-        "place": "Prixod",
-        "pastavchik_id": model.pastavchik_id
-      }
-      await register_supplierModel.create(register);
-    res.status(200).send({
-        error: false,
-        error_code: 200,
-        message: 'Malumotlar tahrirlandi',
-        data: model
-    });
+        
+        await register_supplierModel.create(register);
+        
+        res.status(200).send({
+            error: false,
+            error_code: 200,
+            message: 'Malumotlar tahrirlandi',
+            data: model
+        });
     }
+
     #prixod_table = async(model, prixod_table, insert = true) => {
         if(!insert){
-        await this.#deletePrixodTable(model.id)
-        await this.#deleteRegister_reagent(model.id)
+            await this.#deletePrixodTable(model.id)
+            await this.#deleteRegister_reagent(model.id)
         }
         for(let key of prixod_table){
             var tables = {
@@ -146,21 +155,27 @@ class prixodController {
             await register_reagentModel.create(reagent);
         }
     }
-    #deletePrixodTable = async(doc_id) =>{
-    await prixod_tableModel.destroy({where:{prixod_id: doc_id}})
-    }
-    #deleteRegister_reagent = async(id) => {
-        await register_reagentModel.destroy({where: {doc_id: id}})
-    }
+
     delete = async (req, res, next) => {
-        const model = await prixod_Model.destroy({
-            where:{
+        const model = await prixod_Model.findOne({
+            where: {
                 id: req.params.id
             }
         });
-
+        
+        if (!model) {
+            throw new HttpException(404, "Bunday id mavjud emas");
+        }
+        
+        await prixod_Model.destroy({
+            where: {
+                id: req.params.id
+            }
+        });
+    
         await this.#deletePrixodTable(model.id)
         await this.#deleteRegister_reagent(model.id)
+        await this.#deleteRegister_supplier(model.id)
 
         if(!model){
             throw new HttpException(404, "bunday id yoq")
@@ -173,6 +188,19 @@ class prixodController {
             data: model
         });
     }
+
+    #deletePrixodTable = async (doc_id) =>{
+        await prixod_tableModel.destroy({where:{prixod_id: doc_id},force: true })
+    }
+    
+    #deleteRegister_reagent = async (id) => {
+            await register_reagentModel.destroy({where: {doc_id: id},force: true })
+    }
+
+    #deleteRegister_supplier = async (id) => {
+        await register_supplierModel.destroy({where: {doc_id: id},force: true })
+    }      
+
     checkValidation = (req) => {
         const errors = validationResult(req)
         if (!errors.isEmpty()) {
